requirements.txt
boto3==1.7.45
botocore==1.10.45
docutils==0.14
jmespath==0.9.3
lxml==4.2.2
Pillow==5.1.0
python-dateutil==2.7.3
python-pptx==0.6.10
s3transfer==0.1.13
six==1.11.0
XlsxWriter==1.0.5
